requirements.txt
streamlit
pandas
plotly
openai
openpyxl
sqlalchemy
python-dotenv
chromadb